Repro the problem found in `docker-infrastructure`
After bumping to gradle-release:0.8.0, docker-infrastructure [fails publishing].
The `PublishTest` cannot reproduce it with `withPluginClasspath`, it hides the problem.
But then, the test project has to specify `gradle-release` version (0.9.0 here),
or otherwise it cannot resolve the unnamed version we're currently testing/building.

So this test can repro, but cannot be used to confirm fixes.

The failure in CI and in `PublishTest` looks like this:
```
* What went wrong:
org/jetbrains/kotlin/gradle/dsl/KotlinSingleTargetExtension
> org.jetbrains.kotlin.gradle.dsl.KotlinSingleTargetExtension
```

PS. When minimizing the repro in `Fixtures`, I accidentally discovered
that the order of Gradle plugins matters. The problem goes away when `gradle-release` goes first.
The buildscript/plugin classpath is a fascinating mess.

[fails publishing]: https://github.com/atlassian/docker-infrastructure/actions/runs/9481104563/job/26126902514